@@ -1,14 +1,14 @@
 name: settings_yaml
-version: 2.1.5
+version: 3.0.0-nullsaftey.0
 homepage: https://github.com/bsutton/settings_yaml
 description: Provides a simple means of saving/loading settings from a yaml file without the yaml fuss.
 environment: 
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
 dependencies: 
-  meta: ^1.0.0
-  yaml: ^2.2.0
-  path: ^1.0.0
+  meta: ^1.3.0-nullsafety.6
+  yaml: ^3.0.0-nullsafety.0  
+  path: ^1.8.0-nullsafety.3  
   uuid: ^2.0.4
 dev_dependencies: 
-  test: ^1.0.0
-  pedantic: ^1.9.0
+  test: ^1.16.0-nullsafety.13  
+  pedantic: ^1.10.0-nullsafety.3 
